Updating `copy_files` function of `subtex.py`.
Adding check if destination path already exists. The script will now
through an exception of multiple auxiliary files have the same name.

+++ subtex.py
@@ -283,6 +283,8 @@ def copy_files(list_of_tuples):
     paths_copied = []
     paths_failed = []
     for paths in list_of_tuples:
+        if os.path.exists(paths[1]):
+            raise Exception('Multiple files with name {0}!'.format(paths[1]))
         try:
             shutil.copyfile(paths[0], paths[1])
         except IOError, e:
